Update MessageWindow
Add a parameter to MessageWindow show methods that enables you to specify that the window owner should not be set automatically, that is: leave it at null. This is needed in certain startup scenarios where you want to display a message window, but there is not yet a owner to set.