Update parent/child relation method
Add additional (optional) parameters to set the delete rule and the default column value. The default for the delete rule is Cascade, as this is the default in .Net when it's not specified. When you pass Rule.SetDefault, you can also pass a value that is used for the child column default.