Deferred tool tip
A deferred tool tip is one that is applied directly to the DataGridColumnHeader object after the data grid is loaded. When applied to the DataGridColumnHeader, the tool tip appears in any part of the header; otherwise, it only appears on the actual content of the header (the TextBlock, for instance).